# Fabrication_IHM/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API_Jouje",
    "version": "1.0"
  },
  "paths": {
    "/api/Bois/GetAllBois": {
      "get": {
        "tags": [
          "Bois"
        ],
        "operationId": "GetAllBois",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bois"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bois"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Bois"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Bois/GetBoisById": {
      "get": {
        "tags": [
          "Bois"
        ],
        "operationId": "GetBoisById",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              }
            }
          }
        }
      }
    },
    "/api/Bois/AddBois": {
      "post": {
        "tags": [
          "Bois"
        ],
        "operationId": "AddBois",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bois"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Bois"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Bois"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              }
            }
          }
        }
      }
    },
    "/api/Bois/UpdateBoisObj": {
      "put": {
        "tags": [
          "Bois"
        ],
        "operationId": "UpdateBoisObj",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Bois"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Bois"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Bois"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              }
            }
          }
        }
      }
    },
    "/api/Bois/UpdateBois": {
      "put": {
        "tags": [
          "Bois"
        ],
        "operationId": "UpdateBois",
        "parameters": [
          {
            "name": "idBois",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "nomBois",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "poidsBois",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "region",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bois"
                }
              }
            }
          }
        }
      }
    },
    "/api/Bois/DeleteBois": {
      "delete": {
        "tags": [
          "Bois"
        ],
        "operationId": "DeleteBois",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Guitare/GetAllGuitares": {
      "get": {
        "tags": [
          "Guitare"
        ],
        "operationId": "GetAllGuitares",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Guitare"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Guitare"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Guitare"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Guitare/GetGuitareByID": {
      "get": {
        "tags": [
          "Guitare"
        ],
        "operationId": "GetGuitareByID",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              }
            }
          }
        }
      }
    },
    "/api/Guitare/AddGuitare": {
      "post": {
        "tags": [
          "Guitare"
        ],
        "operationId": "AddGuitare",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Guitare"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Guitare"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Guitare"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              }
            }
          }
        }
      }
    },
    "/api/Guitare/UpdateGuitare": {
      "put": {
        "tags": [
          "Guitare"
        ],
        "operationId": "UpdateGuitare",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Guitare"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Guitare"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Guitare"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              }
            }
          }
        }
      }
    },
    "/api/Guitare/DeletGuitare": {
      "delete": {
        "tags": [
          "Guitare"
        ],
        "operationId": "DeletGuitare",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Guitare"
                }
              }
            }
          }
        }
      }
    },
    "/api/PickUp/GetAllMicros": {
      "get": {
        "tags": [
          "PickUp"
        ],
        "operationId": "GetAllMicros",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PickUp"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PickUp"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PickUp"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/PickUp/AddMicro": {
      "post": {
        "tags": [
          "PickUp"
        ],
        "operationId": "AddMicro",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickUp"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PickUp"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PickUp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PickUp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickUp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickUp"
                }
              }
            }
          }
        }
      }
    },
    "/api/PickUp/UpdateMicro": {
      "put": {
        "tags": [
          "PickUp"
        ],
        "operationId": "UpdateMicro",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PickUp"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PickUp"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PickUp"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PickUp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickUp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickUp"
                }
              }
            }
          }
        }
      }
    },
    "/api/PickUp/DeleteMicro": {
      "delete": {
        "tags": [
          "PickUp"
        ],
        "operationId": "DeleteMicro",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Vibrato/GetAllVibratos": {
      "get": {
        "tags": [
          "Vibrato"
        ],
        "operationId": "GetAllVibratos",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vibrato"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vibrato"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vibrato"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Vibrato/AddVibrato": {
      "post": {
        "tags": [
          "Vibrato"
        ],
        "operationId": "AddVibrato",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vibrato"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Vibrato"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Vibrato"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Vibrato"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vibrato"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vibrato"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vibrato/UpdateVibrato": {
      "put": {
        "tags": [
          "Vibrato"
        ],
        "operationId": "UpdateVibrato",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vibrato"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Vibrato"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Vibrato"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Vibrato"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vibrato"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vibrato"
                }
              }
            }
          }
        }
      }
    },
    "/api/Vibrato/DeleteVibrato": {
      "delete": {
        "tags": [
          "Vibrato"
        ],
        "operationId": "DeleteVibrato",
        "parameters": [
          {
            "name": "_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/WeatherForecast": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "operationId": "GetWeatherForecast",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeatherForecast"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Bois": {
        "type": "object",
        "properties": {
          "idBois": {
            "type": "integer",
            "format": "int32"
          },
          "nomBois": {
            "type": "string",
            "nullable": true
          },
          "poidsBois": {
            "type": "number",
            "format": "double"
          },
          "région": {
            "type": "string",
            "nullable": true
          },
          "stockBois": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "Guitare": {
        "type": "object",
        "properties": {
          "idGuitare": {
            "type": "integer",
            "format": "int32"
          },
          "nomGuitare": {
            "type": "string",
            "nullable": true
          },
          "idClient": {
            "type": "integer",
            "format": "int32"
          },
          "idMicro": {
            "type": "integer",
            "format": "int32"
          },
          "idMicro_1": {
            "type": "integer",
            "format": "int32"
          },
          "idMicro_2": {
            "type": "integer",
            "format": "int32"
          },
          "idBois": {
            "type": "integer",
            "format": "int32"
          },
          "idBois_1": {
            "type": "integer",
            "format": "int32"
          },
          "idBois_2": {
            "type": "integer",
            "format": "int32"
          },
          "idVibrato": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "PickUp": {
        "type": "object",
        "properties": {
          "idMicro": {
            "type": "integer",
            "format": "int32"
          },
          "nomMicro": {
            "type": "string",
            "nullable": true
          },
          "caracteristiquesMicro": {
            "type": "string",
            "nullable": true
          },
          "stockMicro": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Vibrato": {
        "type": "object",
        "properties": {
          "idVibrato": {
            "type": "integer",
            "format": "int32"
          },
          "nomVibrato": {
            "type": "string",
            "nullable": true
          },
          "typeVibrato": {
            "type": "string",
            "nullable": true
          },
          "stockVibrato": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "WeatherForecast": {
        "type": "object",
        "properties": {
          "date": {
            "$ref": "#/components/schemas/DateOnly"
          },
          "temperatureC": {
            "type": "integer",
            "format": "int32"
          },
          "temperatureF": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}